+++ mods/src/title-screen/title-screen-internal.h
@@ -1,3 +1,5 @@
 #pragma once
 
 __attribute__((visibility("internal"))) void _init_welcome();
+__attribute__((visibility("internal"))) extern int version_text_bottom;
+__attribute__((visibility("internal"))) void _init_splashes();